+++ vscode_extension/src/constants.ts
@@ -46,6 +46,7 @@ export namespace PROJECTS {
     export const RUN_EXECUTABLE = `${PROJECTS.CONFIG.KEY}.runExecutable`;
     export const CONFIGURE_OR_CREATE_INI = `${PROJECTS.CONFIG.KEY}.configureOrCreateIni`;
     export const SET_START_PARAMETERS = `${PROJECTS.CONFIG.KEY}.setStartParameters`;
+    export const SET_HOST_APPLICATION = `${PROJECTS.CONFIG.KEY}.setHostApplication`;
     export const SELECT_GROUP_PROJECT = `${PROJECTS.CONFIG.KEY}.pickGroupProject`;
     export const UNLOAD_GROUP_PROJECT = `${PROJECTS.CONFIG.KEY}.unloadGroupProject`;
     export const SELECT_COMPILER = `${PROJECTS.CONFIG.KEY}.selectCompilerConfiguration`;